pyproject.toml
[build-system]
requires = ["poetry >= 2.1.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "seed-vc"
version = "20250701"
description = "Seed VC - Zero-shot voice conversion"
readme = 'README.md'
license = "GPL-3.0"
repository = "https://github.com/hdwlab/custom-seed-vc.git"
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [{include = "seed_vc"}]

[tool.poetry.dependencies]
python = "^3.10,<3.13"
numpy = "^1.26.4"
torch = [
    { version = "~2.7.1", source = "pypi", markers = "sys_platform == 'darwin'" },
    { version = "~2.7.1", source = "torch_cu118", markers = "sys_platform == 'linux'" },
]
torchvision = [
    { version = "~0.22.1", source = "pypi", markers = "sys_platform == 'darwin'" },
    { version = "~0.22.1", source = "torch_cu118", markers = "sys_platform == 'linux'" },
]
torchaudio = [
    { version = "~2.7.1", source = "pypi", markers = "sys_platform == 'darwin'" },
    { version = "~2.7.1", source = "torch_cu118", markers = "sys_platform == 'linux'" },
]
scipy = "~1.13.1"
librosa = "~0.10.2"
huggingface-hub = ">=0.28.1"
munch = "~4.0.0"
einops = "~0.8.0"
gradio = "~5.23.0"
pydub = "~0.25.1"
jiwer = "~3.0.3"
transformers = "~4.46.3"
FreeSimpleGUI = "~5.1.1"
funasr = "~1.2.6"
hydra-core = "~1.3.2"
soundfile = "^0.12.1"
sounddevice = "^0.5.0"
resemblyzer = "^0.1.4"
pyyaml = "^6.0.2"
python-dotenv = "^1.1.0"
python-socketio = "^5.13.0"
eventlet = "^0.40.0"
matplotlib = "^3.10.3"
descript-audio-codec = "~1.0.0"
descript-audiotools = { git = "https://github.com/descriptinc/audiotools" }
onnx = "^1.18.0"
onnxruntime = "^1.21.1"
tensorboard = "^2.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ruff = "0.11.11"
pdbpp = "^0.11.6"

[tool.poetry.group.client.dependencies]
numpy = "^1.26.4"
python-socketio = "^5.13.0"
sounddevice = "^0.5.2"
websocket-client = "^1.8.0"
pydantic = "^2.11.0"
requests = "^2.32.4"

[[tool.poetry.source]]
name = "torch_cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[[tool.poetry.source]]
name = "pypi"
priority = "primary"

[tool.pytest.ini_options]
addopts = "--verbose --durations=0"
testpaths = ["tests"]

[tool.ruff]
exclude = [".pytest_cache", ".venv", "baselines"]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
  "A", # flake8-builtins
  "B", # flake8-bugbear
  "D", # flake8-docstrings,
  "E", # pycodestyle
  "F", # Pyflakes
  "G", # flake8-logging-format
  "I", # isort
  "W", # pycodestyle
  "PT", # flake8-pytest
]
ignore = [
    # [D105] Missing docstring in magic method
    # [D203] `one-blank-line-before-class` (D203) and `no-blank-line-before-class` (D211) are incompatible.
    # [D213] `multi-line-summary-first-line` (D212) and `multi-line-summary-second-line` (D213) are incompatible.
    # [S101] Use of `assert` detected
    # [PT023] pytest-incorrect-mark-parentheses-style
    "D105",
    "D203",
    "D213",
    "S101",
    "PT023",
    # Additional ignores for this project
    "E731",
    "E741",
    "A001",
    "A002",
]


[tool.ruff.lint.per-file-ignores]
"tests/**" = ["INP001"]
"seed_vc/bin/*.py" = ["D100"]
"seed_vc/bin/real-time-gui.py" = ["E402"]
"seed_vc/bin/app*.py" = ["E501"]
"seed_vc/modules/openvoice/*_app.py" = ["E501"]
# NOTE(hdl-members): The following errors are bug of the original code.
"seed_vc/modules/openvoice/modules.py" = ["F821"]
"seed_vc/modules/astral_quantization/bsq.py" = ["F821"]
"seed_vc/modules/bigvgan/alias_free_activation/cuda/activation1d.py" = ["D"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.pydocstyle]
convention = "google"
